+++ controller/OrderController.js
@@ -1,4 +1,5 @@
-// const conn = require('../mariadb');
+const ensureAuthorization = require('../auth'); // 인증모듈
+const jwt = require('jsonwebtoken');
 const mysql = require('mysql2/promise');
 const { StatusCodes } = require('http-status-codes');
 
@@ -12,44 +13,57 @@ const order = async (req, res) => {
         dateStrings : true
     });
 
-    const { items, delivery, totalQuantity, totalPrice, userId, firstBookTitle } = req.body;
+    let authorization = ensureAuthorization(req);
+
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰 입니다."
+        });
+    } else {
+        const { items, delivery, totalQuantity, totalPrice, firstBookTitle } = req.body;
     
-    // delivery 테이블 삽입
-    let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`;
-    let values = [delivery.address, delivery.receiver, delivery.contact];
-    
-    let [results] = await conn.execute(sql, values);
-    let delivery_id = results.insertId;
+        // delivery 테이블 삽입
+        let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?);`;
+        let values = [delivery.address, delivery.receiver, delivery.contact];
 
-    // orders 테이블 삽입
-    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
-                VALUES (?, ?, ?, ?, ?);`;
-    values = [ firstBookTitle, totalQuantity, totalPrice, userId, delivery_id ];
-            
-    [results] = await conn.execute(sql, values);
-    let order_id = results.insertId;
+        let [results] = await conn.execute(sql, values);
+        let delivery_id = results.insertId;
 
-
-    // items를 가지고, 장바구니에서 book_id, quantity 조회
-    sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
-    let [orderItems, fields] = await conn.query(sql, [items]);
+        // orders 테이블 삽입
+        sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
+                    VALUES (?, ?, ?, ?, ?);`;
+        values = [ firstBookTitle, totalQuantity, totalPrice, authorization.id, delivery_id ];
+                
+        [results] = await conn.execute(sql, values);
+        let order_id = results.insertId;
 
 
-    // orderedBook 테이블 삽입
-    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
+        // items를 가지고, 장바구니에서 book_id, quantity 조회
+        sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
+        let [orderItems, fields] = await conn.query(sql, [items]);
 
-    // items .. 배열 : 요소를 하나씩 꺼내서 (foreach문 돌림) >
-    values = [];
-    orderItems.forEach((item) => {
-        // 2차원 배열로 들어
-        values.push([order_id, item.book_id, item.quantity]);
-    })  
 
-    results = await conn.query(sql, [values]);
+        // orderedBook 테이블 삽입
+        sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?;`;
 
-    let result = await deleteCartItems(conn, items);
+        // items .. 배열 : 요소를 하나씩 꺼내서 (foreach문 돌림) >
+        values = [];
+        orderItems.forEach((item) => {
+            // 2차원 배열로 들어
+            values.push([order_id, item.book_id, item.quantity]);
+        })  
 
-    return res.status(StatusCodes.OK).json(result);
+        results = await conn.query(sql, [values]);
+
+        let result = await deleteCartItems(conn, items);
+
+        return res.status(StatusCodes.OK).json(result);
+    }
+
 };
 
 // 결제된 도서 장바구니 삭제
@@ -63,42 +77,68 @@ const deleteCartItems = async (conn, items) => {
         
 // 주문 목록 조회
 const getOrders = async (req, res) => {
-    const conn = await mysql.createConnection({
-        host : 'localhost',
-        user : 'root',
-        password : 'root',
-        database : 'BookShop',
-        dateStrings : true
-    });
+    let authorization = ensureAuthorization(req, res);
+    
+    // instanceof : 왼쪽 변수가 오른쪽 변수 처럼 생겼어
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰 입니다."
+        });
+    } else {
+        const conn = await mysql.createConnection({
+            host : 'localhost',
+            user : 'root',
+            password : 'root',
+            database : 'BookShop',
+            dateStrings : true
+        });
 
-    let sql = `SELECT orders.id, created_at, address, receiver, contact,
-                book_title, total_quantity, total_price
-                FROM orders LEFT JOIN delivery 
-                ON orders.delivery_id = delivery.id;`
+        let sql = `SELECT orders.id, created_at, address, receiver, contact,
+                    book_title, total_quantity, total_price
+                    FROM orders LEFT JOIN delivery 
+                    ON orders.delivery_id = delivery.id;`
 
-    let [rows, fields] = await conn.query(sql);
-    return res.status(StatusCodes.OK).json(rows)
+        let [rows, fields] = await conn.query(sql);
+        return res.status(StatusCodes.OK).json(rows)
+    }
 };
 
 // 주문 상세 조회
 const getOrderDetail = async (req, res) => {
-    const { id } = req.params;
+    let authorization = ensureAuthorization(req, res);
+    
+    // instanceof : 왼쪽 변수가 오른쪽 변수 처럼 생겼어
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰 입니다."
+        });
+    } else {
+        const orderId = req.params.id;
 
-    const conn = await mysql.createConnection({
-        host : 'localhost',
-        user : 'root',
-        password : 'root',
-        database : 'BookShop',
-        dateStrings : true
-    });
+        const conn = await mysql.createConnection({
+            host : 'localhost',
+            user : 'root',
+            password : 'root',
+            database : 'BookShop',
+            dateStrings : true
+        });
 
-    let sql = `SELECT book_id, title, author, price, quantity
-                FROM orderedBook LEFT JOIN books 
-                ON orderedBook.book_id = books.id
-                WHERE order_id = ?;`
+        let sql = `SELECT book_id, title, author, price, quantity
+                    FROM orderedBook LEFT JOIN books 
+                    ON orderedBook.book_id = books.id
+                    WHERE order_id = ?;`
 
-    let [rows, fields] = await conn.query(sql, [id]);
-    return res.status(StatusCodes.OK).json(rows)
+        let [rows, fields] = await conn.query(sql, [orderId]);
+        return res.status(StatusCodes.OK).json(rows)
+    }
 };
 
 module.exports = {
